refactor: create DefaultProxyMetrics::new() method
